requirements.txt
anthropic>=0.40.0
python-telegram-bot>=21.0
fastapi>=0.115.0
uvicorn>=0.32.0
httpx>=0.27.0
python-dotenv>=1.0.0
aiosqlite>=0.20.0
firecrawl-py>=1.0.0
tweepy>=4.14.0
Pillow>=10.0.0
